DDF-4093 Add getByteArray method to ProcessResources
Classes that implemented the ProcessResource interface lacked a good way
to get the resource input stream more than once. This commit adds a
convenience method that lets the user get the input stream data as a
byte array, which will allow them to create as many input streams from
it as they like.